--- src/main/java/com/SchoolManagment/SchoolManagment/ClassRoom/ClassRoomRepository.java
@@ -1,8 +1,0 @@
-package com.SchoolManagment.SchoolManagment.ClassRoom;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ClassRoomRepository extends JpaRepository<ClassRoom,Long> {
-}
