+++ src/main/java/com/SchoolManagment/SchoolManagment/Controller/StudentCourseController.java
@@ -1,0 +1,14 @@
+package com.SchoolManagment.SchoolManagment.Controller;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class StudentCourseController {
+    @Autowired
+    StudentCourseService studentCourseService;
+
+
+
+
+}
