+++ src/main/java/com/SchoolManagment/SchoolManagment/Entity/ClassRoom.java
@@ -1,9 +1,5 @@
-package com.SchoolManagment.SchoolManagment.ClassRoom;
+package com.SchoolManagment.SchoolManagment.Entity;
 
-import com.SchoolManagment.SchoolManagment.BaseEntity.BaseEntity;
-import com.SchoolManagment.SchoolManagment.Course.Course;
-import com.SchoolManagment.SchoolManagment.Teacher.Teacher;
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -21,11 +17,15 @@ import java.util.Set;
 public class ClassRoom extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long classId;
-    private float size;
+    private Long id;
+
+    private Float area;
+
     private String classCode;
+
     @OneToMany(mappedBy = "classRoom")
-    private Set<Course> courses = new HashSet<>();
+    private Set<Course> courses;
+
     @ManyToOne
     private Teacher teacher;
