+++ src/main/java/com/SchoolManagment/SchoolManagment/Entity/Course.java
@@ -1,10 +1,5 @@
-package com.SchoolManagment.SchoolManagment.Course;
+package com.SchoolManagment.SchoolManagment.Entity;
 
-import com.SchoolManagment.SchoolManagment.BaseEntity.BaseEntity;
-import com.SchoolManagment.SchoolManagment.ClassRoom.ClassRoom;
-import com.SchoolManagment.SchoolManagment.StudentCourse.StudentCourse;
-import com.SchoolManagment.SchoolManagment.Teacher.Teacher;
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -22,17 +17,24 @@ import java.util.Set;
 public class Course extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long courseId;
+    private Long id;
+
     private String courseName;
+
     private String description;
+
     private String courseCode;
+
     @ManyToOne
-    @JoinColumn(name = "classRoomId")
+    @JoinColumn(name = "id")
     ClassRoom classRoom;
+
     @OneToOne
     Teacher teacher;
+
     @OneToMany
-    private Set<StudentCourse> studentCourses = new HashSet<>();
+    private Set<StudentCourse> studentCourses;
+
 
 //    public void assignClassRoom(ClassRoom classRoom) {
 //        this.classRoom = classRoom;
