+++ src/main/java/com/SchoolManagment/SchoolManagment/Entity/Student.java
@@ -1,8 +1,5 @@
-package com.SchoolManagment.SchoolManagment.Student;
+package com.SchoolManagment.SchoolManagment.Entity;
 
-import com.SchoolManagment.SchoolManagment.BaseEntity.BaseEntity;
-import com.SchoolManagment.SchoolManagment.StudentCourse.StudentCourse;
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -20,13 +17,18 @@ import java.util.Set;
 public class Student extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long studentId;
+    private Long id;
+
     private String studentName;
-    private int age;
+
+    private Integer age;
+
     private String email;
+
     private String gender;
+
     @OneToMany(mappedBy = "enrolledStudents")
-    private Set<StudentCourse> studentCourses = new HashSet<>();
+    private Set<StudentCourse> studentCourses;
 
 
 }
