+++ src/main/java/com/SchoolManagment/SchoolManagment/Entity/StudentCourse.java
@@ -1,9 +1,8 @@
-package com.SchoolManagment.SchoolManagment.StudentCourse;
+package com.SchoolManagment.SchoolManagment.Entity;
 
-import com.SchoolManagment.SchoolManagment.BaseEntity.BaseEntity;
-import com.SchoolManagment.SchoolManagment.Course.Course;
-import com.SchoolManagment.SchoolManagment.Student.Student;
-import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.SchoolManagment.SchoolManagment.Entity.BaseEntity;
+import com.SchoolManagment.SchoolManagment.Entity.Course;
+import com.SchoolManagment.SchoolManagment.Entity.Student;
 import lombok.*;
 
 import javax.persistence.Entity;
@@ -17,10 +16,13 @@ import javax.persistence.*;
 public class StudentCourse extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long studentCourseId;
-    private float grade;
+    private Long id;
+
+    private Float grade;
+
     @ManyToOne
     private Course enrolledCourses;
+
     @ManyToOne
     private Student enrolledStudents;
