+++ src/main/java/com/SchoolManagment/SchoolManagment/Repository/ClassRoomRepository.java
@@ -1,0 +1,9 @@
+package com.SchoolManagment.SchoolManagment.Repository;
+
+import com.SchoolManagment.SchoolManagment.Entity.ClassRoom;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ClassRoomRepository extends JpaRepository<ClassRoom, Long> {
+}
