+++ src/main/java/com/SchoolManagment/SchoolManagment/Repository/CourseRepository.java
@@ -1,5 +1,6 @@
-package com.SchoolManagment.SchoolManagment.Course;
+package com.SchoolManagment.SchoolManagment.Repository;
 
+import com.SchoolManagment.SchoolManagment.Entity.Course;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
