+++ src/main/java/com/SchoolManagment/SchoolManagment/Repository/StudentRepo.java
@@ -1,5 +1,6 @@
-package com.SchoolManagment.SchoolManagment.Student;
+package com.SchoolManagment.SchoolManagment.Repository;
 
+import com.SchoolManagment.SchoolManagment.Entity.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
