+++ src/main/java/com/SchoolManagment/SchoolManagment/Repository/TeacherRepo.java
@@ -1,5 +1,6 @@
-package com.SchoolManagment.SchoolManagment.Teacher;
+package com.SchoolManagment.SchoolManagment.Repository;
 
+import com.SchoolManagment.SchoolManagment.Entity.Teacher;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
