+++ src/main/java/com/SchoolManagment/SchoolManagment/Service/ClassRoomService.java
@@ -1,9 +1,10 @@
-package com.SchoolManagment.SchoolManagment.ClassRoom;
+package com.SchoolManagment.SchoolManagment.Service;
 
+import com.SchoolManagment.SchoolManagment.Repository.ClassRoomRepository;
+import com.SchoolManagment.SchoolManagment.Entity.ClassRoom;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.security.PublicKey;
 import java.util.List;
 
 @Service
@@ -14,12 +15,15 @@ public class ClassRoomService {
     public List<ClassRoom> getAllClasses() {
         return classRepository.findAll();
     }
+
     public ClassRoom findById(Long classId) {
         return classRepository.findById(classId).get();
     }
+
     public ClassRoom addClass(ClassRoom cls) {
         return classRepository.save(cls);
     }
+
     public void deleteClassRoom(Long classId) {
         classRepository.deleteById(classId);
     }
