+++ src/main/java/com/SchoolManagment/SchoolManagment/Service/StudentCourseService.java
@@ -1,5 +1,7 @@
-package com.SchoolManagment.SchoolManagment.StudentCourse;
+package com.SchoolManagment.SchoolManagment.Service;
 
+import com.SchoolManagment.SchoolManagment.Entity.StudentCourse;
+import com.SchoolManagment.SchoolManagment.Repository.StudentCourseRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
