--- src/main/java/com/SchoolManagment/SchoolManagment/StudentCourse/StudentCourseController.java
@@ -1,18 +1,0 @@
-package com.SchoolManagment.SchoolManagment.StudentCourse;
-
-import com.SchoolManagment.SchoolManagment.Course.Course;
-import com.SchoolManagment.SchoolManagment.Student.Student;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class StudentCourseController {
-    @Autowired
-    StudentCourseService studentCourseService;
-
-
-
-
-}
