--- src/main/java/com/SchoolManagment/SchoolManagment/StudentCourse/StudentCourseRepository.java
@@ -1,6 +1,0 @@
-package com.SchoolManagment.SchoolManagment.StudentCourse;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface StudentCourseRepository extends JpaRepository<StudentCourse,Long> {
-}
