+++ src/main/resources/application.properties
@@ -2,6 +2,6 @@ spring.datasource.url=jdbc:sqlserver://localhost:1433;databaseName=SchoolManagem
 spring.datasource.username=sa
 spring.datasource.password=root
 spring.jpa.show-sql=true
-spring.jpa.hibernate.ddl-auto=update
+spring.jpa.hibernate.ddl-auto=create
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.SQLServer2012Dialect
 spring.datasource.driverClassName=com.microsoft.sqlserver.jdbc.SQLServerDriver
